Backend - Setup routes for test

+++ backend/src/cli/server.rs
@@ -5,7 +5,7 @@ pub struct ServerTypeArgs {
 
     #[arg(short,long)]
     /// Set the port where the port is listening
-    pub port: u64
+    pub port: u16
     
 }
